Change PopularCreator from models/common to datas/user, Add 'uid' to PopularCreator
