+++ src/interfaces/datas/user.ts
@@ -6,3 +6,13 @@ export type UserData = Omit<UserModel, "ownProjectIds" | "receivedCommentIds"> &
     ownProjects: ProjectModel[];
     receivedComments: CommentData[];
 };
+
+export interface PopularCreator {
+    uid: string;
+    firstName: string;
+    lastName: string;
+    profileImageUrl?: string;
+    email: string;
+    totalProjectSuccess: number;
+    totalSupporter: number;
+}
