+++ src/interfaces/models/common.ts
@@ -4,15 +4,6 @@ interface Stage {
     goalFunding: number;
 }
 
-export interface PopularCreator {
-    firstName: string;
-    lastName: string;
-    profileImageUrl?: string;
-    email: string;
-    totalProjectSuccess: number;
-    totalSupporter: number;
-}
-
 export interface Address {
     country: string;
     city: string;
